requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
maturin>=1.0,<2.0
SQLAlchemy==2.0.47
python-dotenv==1.1.1
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pydantic>=2.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
jinja2>=3.1.0
python-magic>=0.4.27
Pillow>=1.1.7
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
argon2-cffi>=23.1.0